+++ packages/zefyr/lib/src/widgets/editor_selection_delegate_mixin.dart
@@ -1,6 +1,7 @@
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/services.dart';
 import 'package:validators/validators.dart';
+import 'package:zefyr/src/widgets/text_selection_controls.dart';
 import 'package:zefyr/zefyr.dart';
 
 import 'editor.dart';
